Add distributed model executor abstraction (#3191)
